uploading_code/requirements.txt
torch==2.0.0+cu118 --index-url https://download.pytorch.org/whl/cu118
torchvision>=0.3
numpy
opencv-python
tqdm>=4.36.0
matplotlib
plotly
scipy
h5py
pycolmap>=0.6.0
kornia>=0.6.11
gdown
lightglue @ git+https://github.com/cvg/LightGlue
